Flytt restklasser til egne filer og initier restutvidetbehandling i fagsakservice
