+++ src/main/kotlin/no/nav/familie/ba/sak/behandling/restDomene/RestPersonResultat.kt
@@ -1,0 +1,43 @@
+package no.nav.familie.ba.sak.behandling.restDomene
+
+import no.nav.familie.ba.sak.behandling.vilkår.PersonResultat
+import no.nav.familie.ba.sak.behandling.vilkår.Vilkår
+import no.nav.familie.ba.sak.nare.Resultat
+import java.time.LocalDate
+import java.time.LocalDateTime
+
+data class RestPersonResultat(
+        val personIdent: String,
+        val vilkårResultater: List<RestVilkårResultat>
+)
+
+data class RestVilkårResultat(
+        val id: Long,
+        val vilkårType: Vilkår,
+        val resultat: Resultat,
+        val periodeFom: LocalDate?,
+        val periodeTom: LocalDate?,
+        val begrunnelse: String,
+        val endretAv: String,
+        val endretTidspunkt: LocalDateTime,
+        val behandlingId: Long,
+        val erVurdert: Boolean? = null
+)
+
+
+fun PersonResultat.tilRestPersonResultat() =
+        RestPersonResultat(personIdent = this.personIdent,
+                           vilkårResultater = this.vilkårResultater.map { vilkårResultat ->
+                               RestVilkårResultat(
+                                       resultat = vilkårResultat.resultat,
+                                       id = vilkårResultat.id,
+                                       vilkårType = vilkårResultat.vilkårType,
+                                       periodeFom = vilkårResultat.periodeFom,
+                                       periodeTom = vilkårResultat.periodeTom,
+                                       begrunnelse = vilkårResultat.begrunnelse,
+                                       endretAv = vilkårResultat.endretAv,
+                                       endretTidspunkt = vilkårResultat.endretTidspunkt,
+                                       behandlingId = vilkårResultat.behandlingId,
+                                       erVurdert = vilkårResultat.resultat != Resultat.IKKE_VURDERT || vilkårResultat.versjon > 0
+                               )
+                           })
